Add option to balance words when lesson contains rare letters
The lesson 'yz' in Italian will likely pick only words with 'z' and zero words with 'y' given that are very rare. This fix allows to pick also all the possible words with 'y'. The lesson may still not be balanced but at least some words with 'y' will appear (possibly repeated).